+++ CHANGELOG.md
@@ -7,8 +7,12 @@ and this project adheres to [Semantic Versioning][SEMVER].
 
 ## [Unreleased]
 
+### Added
+- `post` variant for use with listeners
+
 ### Fixed
 - Responses now work as intended
+
 ## [0.2.1] - 2019-02-16
 
 ### Added